+++ Sources/StreamChat/Controllers/ChannelController/ChannelController.swift
@@ -407,11 +407,7 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
         }
 
         guard !hasLoadedAllPreviousMessages && !isLoadingPreviousMessages else {
-            log.debug("""
-                Skipped loading more previous messages.
-                    - hasLoadedAllPreviousMessages:\(hasLoadedAllPreviousMessages)
-                    - isLoadingPreviousMessages: \(isLoadingPreviousMessages)
-            """)
+            completion?(nil)
             return
         }
 
@@ -465,11 +461,7 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
         }
 
         guard !hasLoadedAllNextMessages && !isLoadingNextMessages else {
-            log.debug("""
-                Skipped loading more next messages.
-                    - hasLoadedAllNextMessages:\(hasLoadedAllNextMessages)
-                    - isLoadingNextMessages: \(isLoadingNextMessages)
-            """)
+            completion?(nil)
             return
         }
 
@@ -518,10 +510,7 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
         }
 
         guard !isLoadingMiddleMessages else {
-            log.debug("""
-                Skipped loading middle messages.
-                    - isLoadingMiddleMessages: \(isLoadingMiddleMessages)
-            """)
+            completion?(nil)
             return
         }